Refactor Speedometer and Fix filtered speed decrement bug (#73)
* refactor(Speedometer): refactor and optimize

* fix: gradual decrement of filtered speed

+++ src/filter/emafilter.cpp
@@ -1,29 +1,29 @@
 #include "emafilter.h"
 
-EmaFilter::EmaFilter(QObject* parent, float alpha)
+EmaFilter::EmaFilter(QObject* parent, double alpha)
     : IFilter(parent), alpha(alpha), oneMinusAlpha(1.0f - alpha), ema(0.0f)
 {}
 
 EmaFilter::~EmaFilter()
 {}
 
-float EmaFilter::getEma() const
+double EmaFilter::getEma() const
 {
     return this->ema;
 }
 
-void EmaFilter::setEma(float value)
+void EmaFilter::setEma(double value)
 {
     this->ema = value;
 }
 
-void EmaFilter::setAlpha(float value)
+void EmaFilter::setAlpha(double value)
 {
     this->alpha = value;
     this->oneMinusAlpha = 1 - value;
 }
 
-float EmaFilter::calculateFilteredOutput(float value)
+double EmaFilter::calculateFilteredOutput(double value)
 {
     this->ema = this->alpha * value + this->oneMinusAlpha * ema;
     return this->ema;
